# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "slim-mom-server",
    "description": "API Documentation"
  },
  "basePath": "/api",
  "paths": {
    "/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register a new user",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                }
              },
              "required": ["username", "password"]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "User registered successfully"
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Authenticate user",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                }
              },
              "required": ["username", "password"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Authentication successful"
          },
          "401": {
            "description": "Authentication failed"
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": ["Auth"],
        "summary": "Logout user",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "refreshToken": {
                  "type": "string"
                }
              },
              "required": ["refreshToken"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Logged out successfully"
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/auth/refresh-token": {
      "post": {
        "tags": ["Auth"],
        "summary": "Generate new access and refresh tokens",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "refreshToken": {
                  "type": "string"
                }
              },
              "required": ["refreshToken"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tokens generated successfully",
            "schema": {
              "type": "object",
              "properties": {
                "accessToken": {
                  "type": "string"
                },
                "refreshToken": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Invalid refresh token"
          }
        }
      }
    },
    "/public/daily-intake": {
      "get": {
        "tags": ["Public"],
        "summary": "Get daily kcal intake and list of non-recommended products",
        "responses": {
          "200": {
            "description": "Successful response",
            "schema": {
              "type": "object",
              "properties": {
                "dailyCalories": {
                  "type": "number"
                },
                "products": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "name": { "type": "string" },
                      "calories": { "type": "number" },
                      "recommended": { "type": "boolean" }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          }
        }
      }
    },
    "/private/daily-intake": {
      "get": {
        "tags": ["Private"],
        "summary": "Get user's daily kcal intake and list of non-recommended products",
        "responses": {
          "200": {
            "description": "Successful response",
            "schema": {
              "type": "object",
              "properties": {
                "dailyIntake": {
                  "type": "number"
                },
                "nonRecommendedProducts": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "name": { "type": "string" },
                      "calories": { "type": "number" },
                      "recommended": { "type": "boolean" }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          }
        }
      }
    },
    "/search": {
      "get": {
        "tags": ["Search"],
        "summary": "Search for products",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "Query string for searching products"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "calories": { "type": "number" },
                  "recommended": { "type": "boolean" }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          }
        }
      }
    },
    "/private/add-consumed-product": {
      "post": {
        "tags": ["Private"],
        "summary": "Add a consumed product",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "productId": { "type": "string" },
                "date": { "type": "string", "format": "date" }
              },
              "required": ["productId", "date"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product added successfully"
          },
          "400": {
            "description": "Invalid request"
          }
        }
      }
    },
    "/private/delete-consumed-product": {
      "delete": {
        "tags": ["Private"],
        "summary": "Delete a consumed product",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "productId": { "type": "string" },
                "date": { "type": "string", "format": "date" }
              },
              "required": ["productId", "date"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product deleted successfully"
          },
          "400": {
            "description": "Invalid request"
          }
        }
      }
    },
    "/private/day-info": {
      "get": {
        "tags": ["Private"],
        "summary": "Get all information for a specific day",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": true,
            "type": "string",
            "format": "date",
            "description": "The date for which to retrieve information"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "product": {
                    "type": "object",
                    "properties": {
                      "name": { "type": "string" },
                      "calories": { "type": "number" },
                      "recommended": { "type": "boolean" }
                    }
                  },
                  "date": {
                    "type": "string",
                    "format": "date-time"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          }
        }
      }
    }
  }
}